(refactor): abstract navigation configs
