+++ src/navigation/index.tsx
@@ -1,0 +1,61 @@
+import { NavigationContainer, ParamListBase } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { FunctionComponent, useCallback, useEffect, useState } from "react";
+import * as SplashScreen from "expo-splash-screen";
+import { Home } from "../pages/Home";
+import { Training } from "../pages/Training";
+import { Workout } from "../pages/Workout";
+import { fetchPoppinsFonts } from "../configs/fonts/poppins";
+
+export interface RootStackParamList extends ParamListBase {
+  Home: undefined;
+  Training: undefined;
+  Workout: undefined;
+}
+
+SplashScreen.preventAutoHideAsync();
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+export const Navigation: FunctionComponent = () => {
+  const [appIsReady, setAppIsReady] = useState(false);
+
+  const onLayoutRootView = useCallback(async () => {
+    if (appIsReady) {
+      await SplashScreen.hideAsync();
+    }
+  }, [appIsReady]);
+
+  useEffect(() => {
+    async function prepare() {
+      try {
+        await fetchPoppinsFonts();
+      } catch (e) {
+        console.warn(e);
+      } finally {
+        setAppIsReady(true);
+      }
+    }
+
+    prepare();
+  }, []);
+
+  if (!appIsReady) {
+    return null;
+  }
+
+  return (
+    <NavigationContainer onReady={onLayoutRootView}>
+      <Stack.Navigator initialRouteName="Home">
+        <Stack.Screen
+          name="Home"
+          component={Home}
+          options={{
+            headerShown: false,
+          }}
+        />
+        <Stack.Screen name="Training" component={Training} />
+        <Stack.Screen name="Workout" component={Workout} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
